refactor: simplify IBC redundant relay check in given restructure of SDK v0.46 (#1288)
* refactor: simplify ibc redundancy check used as sdk middleware

Instead of having the function checkRedundancy check if the call is on CheckTx or SimulateTx, only call the function on CheckTx or SimulateTx

* add godoc
